fix: re-introduce plugin specific templates
In contrast to the previous attempt, explicitly list all plugins, that have NOT been migrated to the new template so far. This way, the list will get smaller and empty over time. This is easily possible here, because this `guiPluginNew.xhtml` is only used by plugins, that have been migrated to the Maven structure so far.